Revert "refactor(onboard): extract dashboard forward lifecycle helpers"
This reverts commit c79cf0dea7bd61f13e79a549976bdcbee6199d0f.
